+++ Compare_Latest/PrivateFrameworks/VectorKit/VKBackgroundRasterTileSource.h
@@ -11,7 +11,7 @@ __attribute__((visibility("hidden")))
 {
 }
 
-- (unsigned int)mapLayerForZoomLevelRange;
+- (unsigned long long)mapLayerForZoomLevelRange;
 - (int)defaultMaximumZoomLevel;
 - (BOOL)maximumZoomLevelBoundsCamera;
