+++ Compare_Latest/PrivateFrameworks/VectorKit/VKLabelModelDelegate-Protocol.h
@@ -8,6 +8,6 @@
 
 @protocol VKLabelModelDelegate
 - (void)labelModel:(VKLabelModel *)arg1 selectedLabelMarkerWillDisappear:(VKLabelMarker *)arg2;
-- (VKMapTile *)labelModel:(VKLabelModel *)arg1 mapTileForTile:(VKTile *)arg2 layer:(unsigned int)arg3;
+- (VKMapTile *)labelModel:(VKLabelModel *)arg1 mapTileForTile:(VKTile *)arg2 layer:(unsigned long long)arg3;
 @end
