+++ Compare_Latest/PrivateFrameworks/VectorKit/VKMapLayer-Protocol.h
@@ -7,6 +7,6 @@
 #import "NSObject.h"
 
 @protocol VKMapLayer <NSObject>
-- (unsigned int)mapLayerPosition;
+- (unsigned long long)mapLayerPosition;
 @end
